+++ supervisely/nn/inference/__init__.py
@@ -26,4 +26,5 @@ from supervisely.nn.inference.tracking.bbox_tracking import BBoxTracking
 from supervisely.nn.inference.tracking.mask_tracking import MaskTracking
 from supervisely.nn.inference.tracking.object_tracking_3d import ObjectTracking3D
 from supervisely.nn.inference.object_detection_3d import ObjectDetection3D
-from supervisely.nn.inference.info import TaskType, CheckpointInfo, DeployInfo, RuntimeType
+from supervisely.nn.inference.utils import CheckpointInfo, DeployInfo, RuntimeType, ModelSource
+from supervisely.nn.task_type import TaskType
