+++ supervisely/nn/task_type.py
@@ -1,0 +1,11 @@
+class TaskType:
+    OBJECT_DETECTION = "object detection"
+    INSTANCE_SEGMENTATION = "instance segmentation"
+    SEMANTIC_SEGMENTATION = "semantic segmentation"
+    POSE_ESTIMATION = "pose estimation"
+    SALIENT_OBJECT_SEGMENTATION = "salient object segmentation"
+    PROMPT_BASED_OBJECT_DETECTION = "prompt-based object detection"
+    INTERACTIVE_SEGMENTATION = "interactive segmentation"
+    PROMPTABLE_SEGMENTATION = "promptable segmentation"
+    TRACKING = "tracking"
+    OBJECT_DETECTION_3D = "object detection 3d"
